integrations/google_slides/requirements.txt
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0
google-auth==2.23.4
Pillow==10.1.0
reportlab==4.0.7